ver1.02 ファイルやディレクトリの構造を変更
