+++ README.md
@@ -20,3 +20,19 @@ ver1.01リリース
 ・>出力文字に色を付けてgrepしやすくなりました
 
 !- Regexオプションを付けずに正規表現を検索ワードにするとバグります。解決方法分からんので放置します。ぜってぇに許さねぇ、バックスラッシュぅぅ！！
+＝＞　2019/7/18 修正しました
+
+ver1.02リリース
+・>ファイル構造を意識して処理を分けました。結構まともな形になった気がします。機能としては特に変わりなく
+・>'\w+'で検索した際にRegexオプションが無ければ、\W+という文字列で検索したかったのですが、処理上無理があると判断したため、実装は見送りします。諦めたことで無理な実装はする必要がなくなったので、バグは修正されました
+正規表現検索の場合は、''を付けないと認識されないため、検索の際には付けるようにお願いします
+
+test
+
+[a-z]+
+\w+
+[\w]+
+\t
+w
+\w
+w+
