+++ bin/hw.rb
@@ -1,0 +1,3 @@
+puts "hello, world!"
+# テストファイル
+# bin以下にはコマンド風に扱えるように設定出来たらいいなあ（願望
